Improved aspect_ratio calculation for the chart wheel
Improved aspect_ratio calculation for the chart wheel: now the aspect ratio takes account the pixels of the character according to the selected monospaced font used by the terminal;
Aspect ratio of the chart wheel will be updated on terminal resize;
Changed limit for nuanced temperament: the second (or the third) temperament will count in the final result if it is 0.6666 of the 1st temperament and not 2/3 exactly. This adjustment makes the calculation engine not discard a temperament for an irrelevant decimal number;
Update help files;
New version released.

+++ src/temperament.c
@@ -1092,9 +1092,9 @@ void abrir_janela_interpretacao_temperamento(ScoreTemperament score, ItemTempera
        NOVO: IDENTIFICAÇÃO DINÂMICA DE INFLUENCIAS SECUNDÁRIAS (70% do Líder)
        ========================================================================= */
     
-    float proporcao_minima = TEMPERAMENT_RANK_PROPORTION;
+    double proporcao_minima = TEMPERAMENT_RANK_PROPORTION;
     // O corte é calculado com base no primeiro colocado da lista (o de maior porcentagem)
-    float corte_significativo = lista[0].porcentagem * proporcao_minima;
+    double corte_significativo = lista[0].porcentagem * proporcao_minima;
     
     int exibiu_conector = 0;
